modify default backoff strategy
